+++ Models/GradeBand.cs
@@ -9,8 +9,8 @@ namespace ARIS1.Models
         public int SubjectId { get; set; }
         public Subject Subject { get; set; } = null!;
 
-        public float MinPercentage { get; set; } // e.g., 80
-        public float MaxPercentage { get; set; } // e.g., 100
+        public decimal MinPercentage { get; set; } // e.g., 80
+        public decimal MaxPercentage { get; set; } // e.g., 100
         public int APSLevel { get; set; } // 7, 6, 5, 4, 3, 2, 0
         public string Grade { get; set; } = string.Empty; // Optional: "A", "B", "C", etc.
